+++ R/dfidx-package.R
@@ -1,33 +1,23 @@
 #' dfidx package: linear models for panel data
-#' 
-#' plm is a package for R which intends to make the estimation of linear panel
-#' models straightforward. plm provides functions to estimate a wide variety of
-#' models and to make (robust) inference.
-#' 
-#' For a gentle and comprehensive introduction to the package, please see the
-#' package's vignette.
-#' 
-#' The main functions to estimate models are:
 #'
-#' - `plm` : panel data estimators using `lm` on transformed data,
-#' -`pgmm` generalized method of moments (GMM) estimation for panel
-#' data,
-#' - `pvcm` variable coefficients models
-#' - `pmg` mean groups (MG), demeaned MG and common correlated effects
-#' (CCEMG) estimators.
+#' ## Overview
 #'
-#' Next to the model estimation functions, the package offers several
-#' functions for statistical tests related to panel data/models.
+#' `dfidx` is a small `R` package which is intended to deal
+#' efficiently with data frames when each observation is identified by
+#' a combination of values of two variables (indexes). `dfidx` exports
+#' the function `dfidx` which takes as main argument a data frame and
+#' returns a `dfidx` object and the `idx`/`idx_name` functions which
+#' respectively return one or all of the indexes and their
+#' names. Methods are provided for `R`'s extractors and for some of
+#' `dplyr`'s verbs.
 #' 
-#' Multiple functions for (robust) variance--covariance matrices are
-#' at hand as well.
+#' ## Installation
 #' 
-#' The package also provides data sets to demonstrate functions and to
-#' replicate some text book/paper results.  Use
-#' `data(package="plm")` to view a list of available data sets in
-#' the package.
+#' `dfidx` is on `CRAN`. Use: `install.packages("dfidx")`.
+#'
+#' To install the developping version, use:
+#' `remotes::install_github("ycroissant/dfidx")`.
 #' 
 #' @name dfidx-package
 #' @docType package
 #' @keywords package
-